Improvements to data anonymizer tool
